Refactor Srtucture
