+++ src/test/java/com/techatpark/sjson/core/BooleanTest.java
@@ -1,7 +1,6 @@
-package com.techatpark.sjson.core.parser;
+package com.techatpark.sjson.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techatpark.sjson.core.Json;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -14,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 /**
  * Tests for Boolean Parsers.
  */
-class BooleanParserTest {
+class BooleanTest {
 
     final ObjectMapper objectMapper = new ObjectMapper();
