+++ src/test/java/com/techatpark/sjson/core/ObjectTest.java
@@ -1,8 +1,7 @@
-package com.techatpark.sjson.core.parser;
+package com.techatpark.sjson.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.JsonParser;
-import com.techatpark.sjson.core.Json;
 import com.techatpark.sjson.core.util.TestDataProvider;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -16,7 +15,8 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 
-class ArrayParserTest {
+class ObjectTest {
+
     final ObjectMapper objectMapper = new ObjectMapper();
 
 
@@ -31,7 +31,7 @@ class ArrayParserTest {
 
         for (Path path : TestDataProvider.getJSONFiles()) {
             String jsonText = Files.readString(path).trim();
-            if (jsonText.startsWith("[")) {
+            if (jsonText.startsWith("{")) {
                 jsonFilePath.add(path);
             }
         }
@@ -52,4 +52,5 @@ class ArrayParserTest {
                 JsonParser.parseReader(new FileReader(path.toFile())),
                 "Reverse JSON Failed for " + path);
     }
+
 }
