+++ src/test/java/com/techatpark/sjson/core/StringTest.java
@@ -1,7 +1,6 @@
-package com.techatpark.sjson.core.parser;
+package com.techatpark.sjson.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techatpark.sjson.core.Json;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -11,7 +10,7 @@ import java.io.StringReader;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class StringParserTest {
+class StringTest {
 
     final ObjectMapper objectMapper = new ObjectMapper();
